requirements.txt
streamlit
numpy
pillow
scikit-learn==1.3.0
matplotlib